Day 11: Replace my own memoization cache with "cached::proc_macro"
By having a '#[cached]' macro, every call to the function passes
through a cache. If the function is called, a cache is checked
and if the exact arguments (stone, count) are already present
then it'll return those.

So... It basically does the same as my homegrown code did, but
probably better.